requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==2.7.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.32.1
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.2
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
